Rework everything into a scenario framework
Now it's possible to write scenarios that have a setup and teardown
cycle and run them. The setup creates the necessary datastore state and
configures ROM to use it, the teardown resets the datastore state to nil
so that it can be run again.

Between setup and teardown the user is dropped into an interactive
console where they can play with ROM to see how the data structures it
uses behave.